backend/requirements.txt
fastapi~=0.115.14
uvicorn~=0.32.1
pydantic~=2.10.3
sqlmodel~=0.0.24
pyjwt~=2.10.1
starlette~=0.40.0
psycopg2-binary
pandas
numpy~=2.1.3

opencv-python~=4.12.0.88
baidu-aip
chardet
python-dotenv~=1.1.1
dotenv~=0.9.9
pycryptodome~=3.23.0
SQLAlchemy~=2.0.41
Markdown~=3.8.2
requests~=2.32.4
cryptography